Removing old filestore-system files and reorganizing the new package
