--- src/db/index.ts
@@ -1,6 +1,0 @@
-export * from './store-index.entity';
-export * from './archive-index.entity';
-export * from './file-index.entity';
-export * from './group-index.entity';
-export * from './index-entity';
-export * from './index-service';
