--- src/dev.ts
@@ -1,6 +1,0 @@
-import { Store } from './index';
-import { join } from 'path';
-
-
-const store = Store.create('435');
-
