+++ src/file-system/file-base.ts
@@ -1,6 +1,6 @@
-import { IndexEntity } from './db/index.entity';
+import { IndexEntity } from '../db/index.entity';
 import { FileStore } from './file-store';
-import { FileType } from './file-type';
+import { FileType } from '../config/file-type';
 
 
 export class FileBase {
