--- src/index.ts
@@ -1,10 +1,0 @@
-export * from './indexed-file';
-export * from './flat-file';
-export * from './group';
-export * from './archive';
-export * from './store';
-export * from './file-state';
-export * from './util';
-export * from './db';
-export * from './config';
-
