--- src/util/index.ts
@@ -1,2 +1,0 @@
-export const isSet = (variable: any): boolean =>
-    variable !== undefined && variable !== null;
